requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.10.6
python-dotenv==1.0.1
word2number==1.1
aiohttp==3.11.11
psycopg2-binary==2.9.10
